Add closure_convert/base.
Migrate ints to use i32 instead of isize.

+++ lowering/base/src/lib.rs
@@ -144,7 +144,7 @@ impl Type {
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub enum IR {
   Var(Var),
-  Int(isize),
+  Int(i32),
   Fun(Var, Box<Self>),
   App(Box<Self>, Box<Self>),
   TyFun(Kind, Box<Self>),
